requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
filelock==3.12.2
Flask==2.3.2
gunicorn==20.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.1
platformdirs==3.8.0
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.23.1
Werkzeug==2.3.3
